Update nginx config to conform to latest version
The nginx config file format has changed slightly so I've updated it
to work with the latest version: `nginx/1.18.0 (Ubuntu)`.

Update the GUIDE with (pretty poor) instructions to remind me how to do
all this.

Add a step in the github workflow to copy across the latest nginx
config. I'm not sure if this will work automatically, the nginx service
might need to be restarted 🤔.